requirements.txt
discord.py==2.4.0
python-dotenv==1.0.0
feedparser==6.0.11
aiohttp==3.10.11
yt-dlp==2024.11.18
PyNaCl==1.5.0
asyncpg==0.29.0
